Add 'FLOW_DOCKER_DONT_PULL' setting
We've recently changed the resolwe 'register' command to
also pull Docker images after process registration is done,
as this prevents using stale or obsolete Docker images in
tests, as well as in production.

In some cases, this might not be desired, so a new setting
was added to prevent pulling images.  This can be useful
during development, since the developer will probably not
want to pull all the images on every test run.

To use this option, simply set the environment variable
'PIPELINES_TEMPLATE_DOCKER_DONT_PULL' to '1'.